Added Tagging API doc readme + removing samples for unsupported languages + Simplified text
