+++ README.md
@@ -1,23 +1,22 @@
-# Musiio Tagging Code Samples
+# Musiio Tagging
 
-This repository contains code samples for using Musiio API **Tagging** service.
+This repository contains code samples and API documentation for using
+**Musiio Tagging** offerings.
 
-Given input tracks via audio files or YouTube links, the Tagging API will return a list of classifications that the Musiio AI predicts for a particular track, for example, content type, bpm, key, genre, instruments, etc.
+Given input tracks via audio files or YouTube links, the Tagging API
+will return a list of classifications that the Musiio AI predicts for a
+particular track, for example, content type, bpm, key, genre, instruments, etc.
 
 ## Index
 
 - [Directory Structure](#directory-structure)
+- [API Documentation](#api-documentation)
 - [Code Samples](samples)
   * [Postman Collection](samples/postman/README.md)
-  * [PHP Samples](samples/php/README.md)
-  * [JavaScript Samples](samples/js/README.md)
-  * [Python Individual Samples](samples/python-individual/README.md)
-  * [Python Integrated Samples](samples/python-integrated/README.md)
-  * [Java Samples](samples/java/README.md)
-  * [Node.js Samples](samples/node/README.md)
+  * [Javascript Samples](samples/js/README.md)
+  * [Python Samples](samples/python-individual/README.md)
+  * [Python Bulk Tagging Samples](samples/python-integrated/README.md)
 - [Windows Executable GUI](windows/README.md)
-- [API Documentation](#api-documentation)
-
 
 ## Directory Structure
 
@@ -28,15 +27,15 @@ Below is a breakdown of the code sample repository structure:
 ```
 Musiio Tagging Code Sample (samples/)
 |
-|--python-individual/	(to run with single track)
-|--python-integrated/	(to run with bulk upload)
-|--java/    (Java samples)
-|--php/     (PHP samples)
-|--js/      (JavaScript samples)
-|--node/    (Node.js samples)
-|--postman/ (Postman collection template)
+|--postman/             (Postman collection)
+|
+|--js/                  (JavaScript samples)
+|
+|--python-individual/   (Python samples)
+|
+|--python-integrated/   (Python bulk tagging samples)
 ```
 
 ## API Documentation
 
-To see the API specifications, click [here](https://app.musiio.com/api)
+To see the detailed API specifications, click [here](https://app.musiio.com/api)
