+++ samples/js/.gitignore
