--- samples/node/upload-audio-file.js
@@ -1,30 +1,0 @@
-const API_KEY = "";     // API key here
-const AUDIO_PATH = "";  // Audio file path here
-
-var axios = require('axios');
-var FormData = require('form-data');
-var fs = require('fs');
-var data = new FormData();
-data.append('audio', fs.createReadStream(AUDIO_PATH));
-
-var config = {
-  method: 'post',
-  url: 'https://api-us.musiio.com/api/v1/upload/file',
-  auth: {
-    username: API_KEY,
-    password: ""
-  },
-  headers: {
-    ...data.getHeaders()
-  },
-  data: data,
-  maxBodyLength: 2000 // Change the maximum body length in bytes allowable for the request if you want to upload larger file
-};
-
-axios(config)
-  .then(function (response) {
-    console.log(JSON.stringify(response.data));
-  })
-  .catch(function (error) {
-    console.log(error);
-  });
